[feat]: silence cron jobs
I don't want to have emails every morning that some project's certifcate
was or wasn't removed, it is ok to see that when we are debugging
things, but not on daily operation.

More generally we should follow black cockpit approach everywhere:

* if things are ok - no lights are lit, (hence the name is "black"),
  i.e. no emails or alerts.
* if there are errors or warnings - we see warning lights,
  i.e. we receive monitoring alerts and/or emails

Silencing thestd out here, cause there is no way to lower verbosity of
lego client unfortunately.