ADHOC fix improve-compatibility: remove broken cert before creating new
The iprevious implementation just checks if a cert exists or no! if it
doesn't it create one. This caused a problem when the cert existed but
it wasn't valid (expired cert or snakeoil cert).

So the solution is to add the relevant checks and delete the cert if
it's invalid or not generated by Let's encrypt and to generate a new
one.

Also a minor refactoring to extract the loop to main.yaml instead the
certificate.yml.