[fix]: install letsencrypt CA certs
when we use staging API, resulting cert is not added to CAs.
In order to allow server connecting to itself - need to let it trust
itself, e.g. when calling REST API on localhost.

it is not enough for PHP to have just server cert, need to pull all
certs from let's encrypt servers and install them.

DO NOT USE THIS ON PRODUCTION, as those staging certs are not controlled
as good as production certs.